rem images

+++ drizzle/schema.ts
@@ -14,7 +14,7 @@ export const products = pgTable("products", {
   name: varchar("name", { length: 256 }),
   collection: varchar("collection", { length: 256 }),
   image: varchar("image", { length: 256 }),
-  price: integer("price"),
+  price: numeric("price"),
   descriptiontitle: varchar("descriptiontitle", { length: 256 }),
   description: varchar("description"),
   descriptiontitle2: varchar("descriptiontitle2", { length: 256 }),
